refactor: rename reducers and split reducers & selectors to diff files
