+++ src/store/reducers/userReducer.ts
@@ -1,6 +1,6 @@
 import { createReducer } from '@reduxjs/toolkit';
 import {
-  addGameSettings, addUserId, addUserName, addUserRoom,
+  addUserId, addUserName, addUserRoom,
 } from '@/store/actions/userActions';
 
 const initialStateUser = {
@@ -9,13 +9,7 @@ const initialStateUser = {
   room: '',
 };
 
-const initialStateGame = {
-  gameName: '',
-  voteSystem: '',
-  dealerRights: '',
-};
-
-export const userReducer = createReducer(initialStateUser, (builder) => {
+export const user = createReducer(initialStateUser, (builder) => {
   builder
     .addCase(addUserId, (state, action) => {
       state.id = action.payload;
@@ -25,16 +19,6 @@ export const userReducer = createReducer(initialStateUser, (builder) => {
     })
     .addCase(addUserRoom, (state, action) => {
       state.room = action.payload;
-    })
-    .addDefaultCase((state) => state);
-});
-
-export const gameReducer = createReducer(initialStateGame, (builder) => {
-  builder
-    .addCase(addGameSettings, (state, action) => {
-      state.gameName = action.payload.gameName;
-      state.voteSystem = action.payload.voteSystem;
-      state.dealerRights = action.payload.dealerRights;
     })
     .addDefaultCase((state) => state);
 });
