+++ src/store/selectors/gameSettingsSelector.ts
@@ -1,0 +1,4 @@
+import { createSelector } from '@reduxjs/toolkit';
+import { RootState } from '@/store';
+
+export const gameSettingsSelector = createSelector((state: RootState) => state.gameSettings, (state) => state);
