+++ src/store/selectors/userSelectors.ts
@@ -1,12 +1,8 @@
 import { createSelector } from '@reduxjs/toolkit';
 import { RootState } from '@/store';
 
-const user = (state: RootState) => state.userReducer;
-
 export const userSelectors = {
-  userId: createSelector(user, (state) => state.id),
-  userName: createSelector(user, (state) => state.name),
-  userRoom: createSelector(user, (state) => state.room),
+  userId: createSelector((state: RootState) => state.user, (state) => state.id),
+  userName: createSelector((state: RootState) => state.user, (state) => state.name),
+  userRoom: createSelector((state: RootState) => state.user, (state) => state.room),
 };
-
-export const gameSettings = createSelector((state: RootState) => state.gameReducer, (state) => state);
